requirements.txt
Pillow>=8.0.0
opencv-python>=4.5.0
pyinstaller>=5.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
jaraco.text>=3.0.0
jaraco.functools>=3.0.0
jaraco.context>=4.0.0
jaraco.collections>=3.0.0